Refactor store
