+++ src/setup.js
@@ -1,12 +1,12 @@
 import React, { PureComponent } from 'react';
 import { Provider } from 'react-redux';
-import createStore from './store/createStore';
+import store from './store';
 import App from './App';
 
 export default function setup() {
   return class Root extends PureComponent {
     state = {
-      store: createStore(() => this.setState({ loaded: true })),
+      store: store.createStore(() => this.setState({ loaded: true })),
       loaded: false,
     }
